requirements.txt
# ============================================
# MULTIMODAL FINANCIAL RAG - REQUIREMENTS
# ============================================

# --- PDF Processing ---
PyMuPDF==1.24.1
pdf2image==1.16.3
unstructured==0.12.4
pillow==10.2.0

# --- Embeddings ---
sentence-transformers==2.3.1

# --- Vector Database ---
chromadb==0.4.22

# --- LLM & Orchestration ---
langchain==0.1.9
langchain-community==0.0.24
langchain-google-genai
google-generativeai

# --- API ---
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# --- Configuration ---
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# --- Utilities ---
loguru==0.7.2
tqdm==4.66.2
requests==2.31.0

# --- Data Handling ---
pandas==2.2.0
numpy==1.26.4